scraper/requirements.txt
httpx>=0.27.0
supabase>=2.9.0
python-dotenv>=1.0.0
playwright>=1.40.0
